# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/user/all": {
      "get": {
        "operationId": "UserController_getUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "UserController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/create": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reason Code",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Создание пользователя",
                  "value": {
                    "name": "name",
                    "password": "password"
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/edit": {
      "patch": {
        "operationId": "UserController_edit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Создание пользователя",
                  "value": {
                    "id": 1,
                    "name": "name",
                    "password": "password"
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/operation-type/all": {
      "get": {
        "operationId": "OperationTypeController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "operation type"
        ]
      }
    },
    "/operation-type/{id}": {
      "get": {
        "operationId": "OperationTypeController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "operation type"
        ]
      },
      "delete": {
        "operationId": "OperationTypeController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "operation type"
        ]
      }
    },
    "/operation-type/create": {
      "post": {
        "operationId": "OperationTypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reason Code",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Создание типа операции",
                  "value": {
                    "name": "name"
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "operation type"
        ]
      }
    },
    "/operation-type/edit": {
      "patch": {
        "operationId": "OperationTypeController_edit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Редактирование типа операции",
                  "value": {
                    "id": 1,
                    "name": "name"
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "operation type"
        ]
      }
    },
    "/finance-operation": {
      "get": {
        "operationId": "FinanceOperationController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    },
    "/finance-operation/{id}": {
      "get": {
        "operationId": "FinanceOperationController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      },
      "delete": {
        "operationId": "FinanceOperationController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    },
    "/finance-operation/user/{userId}": {
      "get": {
        "operationId": "FinanceOperationController_getByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    },
    "/finance-operation/operation-type/{operationTypeId}": {
      "get": {
        "operationId": "FinanceOperationController_getByOperationType",
        "parameters": [
          {
            "name": "operationTypeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    },
    "/finance-operation/create": {
      "post": {
        "operationId": "FinanceOperationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reason Code",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Создание финансовой операции",
                  "value": {
                    "userId": 1,
                    "operationTypeId": 1,
                    "sum": 1000
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    },
    "/finance-operation/edit": {
      "patch": {
        "operationId": "FinanceOperationController_edit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Reason Code",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "test": {
                  "summary": "Пример",
                  "description": "Редактирование финансовой операции",
                  "value": {
                    "id": 1,
                    "userId": 1,
                    "operationTypeId": 1,
                    "sum": 1000
                  },
                  "externalValue": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "finance operation"
        ]
      }
    }
  },
  "info": {
    "title": "Home finance",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}